Improve popup box keyboard control
Set ticked/selected item as highlighted on menu construction and
prevent fake mouse events from resetting this.

Previously when a menu was opened with the enter key a user would have to press escape key to close the menu. This was inconsistent if the selected/ticked item was highlighted using the arrow keys and then enter pressed.

Also previosuly keyboard navigation would start at the top/bottom of the list depending on whether up/down was pressed but this change uses the ticked/selected item as the start point.